requirements.txt
Django == 2.1.5
django-ckeditor == 5.9.0
django-cleanup == 5.0.0
django-crispy-forms == 1.9.2
django-js-asset == 1.2.2	